Renamed variables and added tests
