+++ save/action.yml
@@ -11,9 +11,6 @@ inputs:
   upload-chunk-size:
     description: 'The chunk size used to split up large files during upload, in bytes'
     required: false
-  restored-key:
-    description: 'Cache key restored from the restore action'
-    required: false
 runs:
   using: 'node16'
   main: '../dist/save-only/index.js'
